Для пункта 2.1 /dict (меню словари)

+++ src/main/kotlin/ru/sber/kotlin/school/telegram/bot/model/Dictionary.kt
@@ -14,7 +14,7 @@ import javax.persistence.Table
 data class Dictionary(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private val id: Int,
+    private val id: Long,
 
     private val name: String,
